requirement.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.7
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
filelock==3.12.4
Flask==3.0.2
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
multidict==6.0.5
mysql==0.0.3
mysqlclient==2.2.4
numpy==1.26.4
panda==0.3.1
pillow==10.2.0
platformdirs==3.11.0
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
runserver==1.0.4
sqlparse==0.4.4
twilio==9.0.2
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.24.5
virtualenvwrapper-win==1.2.7
Werkzeug==3.0.1
XlsxWriter==3.2.0
yarl==1.9.4
